refactor: OrderTable과 TableGroup을 별도 패키지로 분리
