+++ src/main/java/kitchenpos/order/Order.java
@@ -11,6 +11,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
+import kitchenpos.table.OrderTable;
 
 
 @Entity(name = "orders")
