+++ src/main/java/kitchenpos/table/request/TableGroupCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.tablegroup.request;
+package kitchenpos.table.request;
 
 import java.util.List;
