+++ src/test/java/kitchenpos/order/OrderTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.params.provider.EnumSource.Mode.EXCLUDE;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import kitchenpos.table.OrderTable;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Nested;
