+++ src/test/java/kitchenpos/table/application/OrderTableServiceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.application;
+package kitchenpos.table.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -10,13 +10,13 @@ import java.util.List;
 import kitchenpos.order.Order;
 import kitchenpos.order.OrderLineItem;
 import kitchenpos.order.OrderStatus;
-import kitchenpos.order.OrderTable;
 import kitchenpos.order.persistence.OrderRepository;
-import kitchenpos.order.persistence.OrderTableRepository;
 import kitchenpos.order.request.OrderTableCreateRequest;
 import kitchenpos.support.ServiceTest;
-import kitchenpos.tablegroup.TableGroup;
-import kitchenpos.tablegroup.persistence.TableGroupRepository;
+import kitchenpos.table.OrderTable;
+import kitchenpos.table.TableGroup;
+import kitchenpos.table.persistence.OrderTableRepository;
+import kitchenpos.table.persistence.TableGroupRepository;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Nested;
